+++ cli/smartcontract/testdata/rpcbindings/invalid1/invalid.yml
