+++ cli/smartcontract/testdata/rpcbindings/invalid2/invalid.go
@@ -1,7 +1,7 @@
-package invalid5
+package invalid2
 
 import "github.com/nspcc-dev/neo-go/pkg/interop/runtime"
 
 func Main() {
-	runtime.Notify("Non declared event")
+	runtime.Notify("SomeEvent", "p1", "p2")
 }
