+++ cli/smartcontract/testdata/rpcbindings/invalid2/invalid.yml
