+++ cli/smartcontract/testdata/rpcbindings/invalid3/invalid.yml
