+++ cli/smartcontract/testdata/rpcbindings/invalid4/invalid.go
@@ -1,4 +1,4 @@
-package invalid8
+package invalid4
 
 import "github.com/nspcc-dev/neo-go/pkg/interop/runtime"
