+++ cli/smartcontract/testdata/rpcbindings/invalid4/invalid.yml
