+++ cli/smartcontract/testdata/rpcbindings/notifications/config.yml
