+++ cli/smartcontract/testdata/rpcbindings/notifications/config_extended.yml
