+++ cli/smartcontract/testdata/rpcbindings/notifications/config_guessed.yml
