+++ cli/smartcontract/testdata/rpcbindings/notifications/notifications.go
