+++ cli/smartcontract/testdata/rpcbindings/notifications/rpcbindings.out
