+++ cli/smartcontract/testdata/rpcbindings/notifications/rpcbindings_extended.out
