+++ cli/smartcontract/testdata/rpcbindings/notifications/rpcbindings_guessed.out
