+++ cli/smartcontract/testdata/rpcbindings/structs/config.yml
