+++ cli/smartcontract/testdata/rpcbindings/structs/rpcbindings.out
