+++ cli/smartcontract/testdata/rpcbindings/structs/rpcbindings_dynamic_hash.out
