+++ cli/smartcontract/testdata/rpcbindings/structs/structs.go
