+++ cli/smartcontract/testdata/rpcbindings/types/config.yml
