+++ cli/smartcontract/testdata/rpcbindings/types/rpcbindings.out
