+++ cli/smartcontract/testdata/rpcbindings/types/rpcbindings_dynamic_hash.out
