+++ cli/smartcontract/testdata/rpcbindings/types/types.go
